+++ src/new_achitecture/modules/budgetTravelTips/presentation/BTTStartComponent/GalleryDynamicComp/GalleryDynamicComponent.module.css
@@ -1,0 +1,22 @@
+.image {
+    width: 100%;
+}
+
+.sectionTitle {
+    color: #000;
+    font-size: 3rem;
+    text-align: center;
+    padding: 5px;
+
+    margin: 0;
+}
+
+.cityDetailsMain {
+    color: #000;
+    text-align: center;
+    font-size: 1.2rem;
+    text-indent: 0;
+    padding: 20px;
+
+    margin: 0 0 20px 0;
+}
