--- src/new_achitecture/modules/budgetTravelTips/presentation/BTTStartComponent/GalleryMUIComponent.module.css
@@ -1,11 +1,0 @@
-.image {
-    width: 200px;
-    height: 200px;
-}
-
-.imageList {
-    margin-left: 10vw;
-    display: flex;
-    justify-items: center;
-    align-items: center;
-}
