+++ src/new_achitecture/modules/budgetTravelTips/presentation/BTTStartComponent/GalleryVersions/GalleryMUIComponent.module.css
@@ -1,0 +1,11 @@
+.image {
+    width: 200px;
+    height: 200px;
+}
+
+.imageList {
+    margin-left: 10vw;
+    display: flex;
+    justify-items: center;
+    align-items: center;
+}
